Fix warmup_rate

+++ main.py
@@ -85,7 +85,7 @@ if __name__ == '__main__':
     parser.add_argument('--img_ch', type=int, default=3)
     parser.add_argument('--output_ch', type=int, default=1)
     parser.add_argument('--num_epochs', type=int, default=100)
-    parser.add_argument('--warmup_rate', type=int, default=0)
+    parser.add_argument('--warmup_rate', type=float, default=0)
     parser.add_argument('--name', type=str, default='test')
     # parser.add_argument('--num_epochs_decay', type=int, default=70)
     parser.add_argument('--batch_size', type=int, default=1)
